PerlinGrid is a single file like everything else
