+++ src/components/GalleryBackground/animations/PerlinGrid.tsx
@@ -1,9 +1,38 @@
 import P5 from 'p5';
-import { useP5DupeRemover } from '../../../../utils/p5DupeRemover';
+import * as dat from 'dat.gui';
+import { useP5DupeRemover } from '../../../utils/p5DupeRemover';
 import Sketch from 'react-p5';
-import { colours } from '../../../../utils/colourUtils';
-import { createDatGui, settings } from './gui';
+import { colours } from '../../../utils/colourUtils';
+import { removeDatGui } from '../../../utils/datGuiUtils';
 import { useEffect, useState } from 'react';
+
+export const settings = {
+  rotation: false,
+  squircles: false,
+  mouseStrength: 30,
+  colourOffset: 0,
+  speedX: 0.005,
+  speedY: 0.005,
+  speedZ: 0.01,
+}
+
+export function createDatGui() {
+  // Deals with duplicate created due to safe mode in development
+  removeDatGui();
+
+  const gui = new dat.GUI();
+
+  gui.add(settings, 'mouseStrength', 0, 100).name('Mouse Strength');
+  gui.add(settings, 'colourOffset', 0, 360).name('Colour Offset');
+
+  const speedFolder = gui.addFolder('Speed');
+  speedFolder.add(settings, 'speedX', -0.02, 0.02).name('Speed X');
+  speedFolder.add(settings, 'speedY', -0.02, 0.02).name('Speed Y');
+  speedFolder.add(settings, 'speedZ', -0.02, 0.02).name('Speed Z');
+
+  gui.add(settings, 'rotation').name('Rotation');
+  gui.add(settings, 'squircles').name('Squircles');
+}
 
 const ARRAY_SIZE_MAX = 20;
